fix(react): split providers from context modules for React Doctor 100/100
react-doctor 0.7.8 re-weighted `only-export-components`, dropping the app to
97, mobile to 97 and ui to 96. Each finding was a context module exporting its
Provider component alongside hooks/contexts, which blocks Fast Refresh from
preserving state.

Split the component out of every flagged module so each file exports either
components or non-components, never both:

- ui: `theme-provider.tsx`, `border-calculator-provider.tsx`, `toast-context.ts`
- app: `context/recipe-provider.tsx`
- mobile: `meter/use-drag-offset.ts`

This also lets the four existing `only-export-components` suppressions in
recipe-context.tsx go away rather than adding four more.

All four projects back to 100/100.
