+++ apps/dorkroom/src/app/pages/development-recipes/context/index.ts
@@ -4,17 +4,17 @@ export type {
   RecipeDataState,
   RecipeModalsActions,
   RecipeModalsState,
-  RecipeProviderProps,
   RecipeUIState,
 } from './recipe-context';
 export {
   RecipeActionsContext,
   RecipeDataContext,
   RecipeModalsContext,
-  RecipeProvider,
   RecipeUIContext,
   useRecipeActionsContext,
   useRecipeDataContext,
   useRecipeModalsContext,
   useRecipeUIContext,
 } from './recipe-context';
+export type { RecipeProviderProps } from './recipe-provider';
+export { RecipeProvider } from './recipe-provider';
