+++ apps/dorkroom/src/app/pages/development-recipes/context/recipe-context.tsx
@@ -5,7 +5,7 @@ import type {
   FilmdevMappingResult,
 } from '@dorkroom/logic';
 import type { DevelopmentCombinationView } from '@dorkroom/ui';
-import { createContext, type ReactNode, use } from 'react';
+import { createContext, use } from 'react';
 
 /**
  * Modal state for the development recipes page.
@@ -135,7 +135,6 @@ export interface RecipeContextValue {
  * Context for recipe modals state and actions.
  * Separated to allow independent updates.
  */
-// eslint-disable-next-line react-doctor/only-export-components -- dedicated context module: contexts/hooks/provider co-located by design, not a component file with stray exports
 export const RecipeModalsContext = createContext<
   (RecipeModalsState & RecipeModalsActions) | null
 >(null);
@@ -144,21 +143,18 @@ export const RecipeModalsContext = createContext<
  * Context for recipe data.
  * Separated to allow independent updates.
  */
-// eslint-disable-next-line react-doctor/only-export-components -- dedicated context module: contexts/hooks/provider co-located by design, not a component file with stray exports
 export const RecipeDataContext = createContext<RecipeDataState | null>(null);
 
 /**
  * Context for recipe actions.
  * Actions are stable references that don't change often.
  */
-// eslint-disable-next-line react-doctor/only-export-components -- dedicated context module: contexts/hooks/provider co-located by design, not a component file with stray exports
 export const RecipeActionsContext = createContext<RecipeActions | null>(null);
 
 /**
  * Context for UI state.
  * Contains mobile detection, animations, etc.
  */
-// eslint-disable-next-line react-doctor/only-export-components -- dedicated context module: contexts/hooks/provider co-located by design, not a component file with stray exports
 export const RecipeUIContext = createContext<RecipeUIState | null>(null);
 
 /**
@@ -213,47 +209,4 @@ export function useRecipeUIContext() {
     throw new Error('useRecipeUIContext must be used within a RecipeProvider');
   }
   return context;
-}
-
-/**
- * Props for the RecipeProvider component.
- */
-export interface RecipeProviderProps {
-  children: ReactNode;
-  modals: RecipeModalsState & RecipeModalsActions;
-  data: RecipeDataState;
-  actions: RecipeActions;
-  ui: RecipeUIState;
-}
-
-/**
- * Provider component that makes recipe context available to children.
- *
- * This allows incremental adoption: wrap parts of the component tree
- * with this provider to enable context-based access to recipe state.
- *
- * @example
- * ```tsx
- * <RecipeProvider modals={modals} data={data} actions={actions} ui={ui}>
- *   <RecipeModals />
- *   <RecipeResultsSection />
- * </RecipeProvider>
- * ```
- */
-export function RecipeProvider({
-  children,
-  modals,
-  data,
-  actions,
-  ui,
-}: RecipeProviderProps) {
-  return (
-    <RecipeModalsContext value={modals}>
-      <RecipeDataContext value={data}>
-        <RecipeActionsContext value={actions}>
-          <RecipeUIContext value={ui}>{children}</RecipeUIContext>
-        </RecipeActionsContext>
-      </RecipeDataContext>
-    </RecipeModalsContext>
-  );
 }
