+++ packages/ui/src/components/border-calculator/border-calculator-provider.tsx
@@ -1,0 +1,19 @@
+import type { ReactNode } from 'react';
+import {
+  BorderCalculatorContext,
+  type BorderCalculatorContextValue,
+} from './border-calculator-context';
+
+interface BorderCalculatorProviderProps {
+  value: BorderCalculatorContextValue;
+  children: ReactNode;
+}
+
+export function BorderCalculatorProvider({
+  value,
+  children,
+}: BorderCalculatorProviderProps) {
+  return (
+    <BorderCalculatorContext value={value}>{children}</BorderCalculatorContext>
+  );
+}
