+++ packages/ui/src/components/toast-context.ts
@@ -1,0 +1,21 @@
+import { createContext, use } from 'react';
+
+export type ToastType = 'success' | 'error' | 'info';
+
+export interface ToastContextValue {
+  showToast: (message: string, type?: ToastType) => void;
+}
+
+export const ToastContext = createContext<ToastContextValue | null>(null);
+
+export function useOptionalToast(): ToastContextValue | null {
+  return use(ToastContext);
+}
+
+export function useToast(): ToastContextValue {
+  const context = use(ToastContext);
+  if (!context) {
+    throw new Error('useToast must be used within a ToastProvider');
+  }
+  return context;
+}
